\name{addnumbers-package}
\alias{addnumbers-package}
\alias{addnumbers}
\docType{package}
\title{
\packageTitle{addnumbers}
}
\description{
\packageDescription{addnumbers}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{addnumbers}
\packageIndices{addnumbers}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{addnumbers}

Maintainer: \packageMaintainer{addnumbers}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
